fastr_project/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.1.0
